LayoutTests/crypto/subtle/hmac-import-jwk-expected.txt

Test importing a JWK key for HMAC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...

PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'hmac'
PASS key.algorithm.length is 32
PASS key.usages is ["sign", "verify"]

Using the key to sign message 'foo'...
PASS byteArrayToHexString(new Uint8Array(signature)) is '[e0 37 36 fe 09 88 92 b2 a2 da 77 81 24 31 f7 c0 14 d3 2e 2f d6 9f 3b cf f8 83 ac 92 3a 8f a2 da]'

Verifying the signature...
PASS verificationResult is true
PASS successfullyParsed is true

TEST COMPLETE
